chore: add hardcoded favicons back
I would love to hear more context on this one. I'm adding it back since we cant have getServerSideProps and getStaticProps both.

Or we choose wagmi cache or we get a static favicon from CDN.

If we move favicon to `getServerSideProps` we would call the CMS on every page request.